Added GUIEvent concept.
In different places, one may register a callback on a GUIEvent declared
in GUIEvents or trigger the event. When triggered, all registered
callbacks will be called.

This should be useful for GUI events, where one GUI element does not
depend on another, but they both interact with the same event.